first refactoring commit
* nothing works yet, just moved around code, renamed variables, changed code structure and such stuff
* server file still the biggest one, but loads of stuff needs to be exported to other files and MyDuel needs to be merged into Duel class
