--- game.py
@@ -1,7 +1,0 @@
-players = {}
-server = None
-websocket_server = None
-german_db = None
-japanese_db = None
-spanish_db = None
-rebooting = False
