+++ ygo/globals.py
@@ -1,0 +1,8 @@
+german_db = None
+japanese_db = None
+lflist = {}
+rebooting = False
+server = None
+spanish_db = None
+strings = {}
+websocket_server = None
