+++ ygo/websockets.py
@@ -1,10 +1,11 @@
 import logging
 from twisted.internet import reactor
+from twisted.internet import ssl
 from autobahn.twisted.websocket import WebSocketServerFactory, \
      WebSocketServerProtocol
 from gsb import Caller
-import game
 
+import globals
 
 class WSProtocol(WebSocketServerProtocol):
 
@@ -18,12 +19,12 @@ class WSProtocol(WebSocketServerProtocol):
 		self.sendMessage(line)
 
 	def onOpen(self):
-		game.server.on_connect(Caller(self))
+		globals.server.on_connect(Caller(self))
 		self.web = True
-		self.parser = game.server.default_parser
+		self.parser = globals.server.default_parser
 
 	def onConnect(self, request):
-		self.server = game.server
+		self.server = globals.server
 		self.server.connections.append(self)
 		self._parser = None
 		self.encode_args = ('utf-8', 'replace')
@@ -31,9 +32,9 @@ class WSProtocol(WebSocketServerProtocol):
 
 
 	def onClose(self, wasClean, code, reason):
-		if self in game.server.connections:
-			game.server.connections.remove(self)
-			game.server.on_disconnect(Caller(self))
+		if self in globals.server.connections:
+			globals.server.connections.remove(self)
+			globals.server.on_disconnect(Caller(self))
 
 	@property
 	def parser(self):
@@ -53,3 +54,17 @@ class WSProtocol(WebSocketServerProtocol):
 	def notify(self, *args, **kwargs):
 		"""Notify this connection of something."""
 		self.server.notify(self, *args, **kwargs)
+
+def start_websocket_server(port, cert, key):
+	if cert:
+		context_factory = ssl.DefaultOpenSSLContextFactory(key, cert)
+		url = 'wss://0.0.0.0:%d' % port
+	else:
+		context_factory = None
+		url = 'ws://0.0.0.0:%d' % port
+	factory = WebSocketServerFactory(url)
+	factory.protocol = WSProtocol
+	if context_factory:
+		globals.websocket_server = reactor.listenSSL(port, factory, context_factory)
+	else:
+		globals.websocket_server = reactor.listenTCP(port, factory)
